Still integrating Users object and privileges into the bot.
Realized that there will have to some way for the system administrator
to be added as the admin to the bot without difficulty.  I added a
command line argument, -i or —init which will tell the bot to allow the
first user who messages it on IRC privately to become admin with the
INIT command.  Still coding it though.  Going to have to rework the
main some more to organize things better.  Might need to create the new
Script object first.