Added a way to initialize the bot, refactored some of the code in main.
Added a way for the bot to be initialized with its first user, and thus
admin, I think it is a bit clunky, but for now it will work to get
things off the ground.  Also refactored main so that it passes the
Network object to the command parser for private messages and thus
allows it to act directly upon network object without gaming around
with “directives” or some such.  Seems cleaner.