Fixed Issue #23
Snowboard now will choose an alternate nick to use.  It adds a random 5
digit number to its primary nick and will keep trying to do so until
its NICK command is accepted.  There is the potential for a bug here,
if the server is lagged, since the server does not send a confirmation
response for a good nick during the initial connection.